+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## 0.8.1 (?)
+## 0.8.1 (2025-05-15)
+
+- Fixed support for distinct flag in reader stores.
+
+- Added support for `Encoder` expressions in `--encoder` in KIF CLI.
 
 - Fixed handling of `--no-resolved` in KIF CLI.